requirements.txt
opencv_python==4.5.1.48
pydensecrf==1.0rc2
tqdm==4.31.1
numpy==1.20.2
torchvision==0.9.1+cu102
matplotlib==3.2.2
scikit_image==0.16.2
~orch==1.3.0
Pillow==8.3.1
skimage==0.0
tensorboardX==2.4
torch==1.9.0
